Separated out API into separate file
