+++ backend/server.py
@@ -1,15 +1,14 @@
 import threading
 
-from flask import Flask, Response, render_template, send_from_directory, request, abort
+from flask import Flask, Response, render_template, send_from_directory
 
 from streaming.live_streaming import generate
 from streaming.static import generate_static
 from streaming.rtsp import RTSPStreamer
 
 from computer_vision.hog_detection import HOGDetectionModel
-from database.database import DatabaseHelper
-from utils.utils import *
 from incidents.ftp import fetch_todays_incidents
+from api import api, db_helper
 
 rtsp_config = { "rtsp_url" : "rtsp://admin:admin@172.24.28.36/11" }
 streamer = RTSPStreamer(rtsp_config)
@@ -19,7 +18,6 @@ streamer = RTSPStreamer(rtsp_config)
 # DETECTED BY AWS ELASTIC BEANSTALK
 application = Flask(__name__, static_url_path='')
 application.config.from_object('config')  # configure flask server
-db_helper = DatabaseHelper(application)  # initialize database helper
 
 @application.route('/stream', methods=['GET'])
 def stream():
@@ -47,69 +45,14 @@ def ping():
     return "pong"
 
 
-@application.route('/api/cameras', methods=['GET', 'POST'])
-def cameras():
-    if request.method == 'GET':
-        return jsonify_result(db_helper.get_camera())
-    elif request.method == 'POST':
-        if check_body(request, 'url'):
-            url = request.get_json()['url']
-            response = unwrap_db_result(db_helper.add_camera(url))
-            response.status_code = 201;
-            return response
-        abort(400)
-
-
-@application.route('/api/incidents', methods=['GET'])
-def incidents():
-    camera_id = request.args.get('camera_id')
-    if camera_id is not None:
-        return jsonify_result(db_helper.get_incidents_by_camera_id(camera_id))
-    else:
-        return jsonify_result(db_helper.get_incident())
-
-@application.route('/api/object_set', methods=['GET', 'POST'])
-def object_set():
-    if request.method == 'GET':
-        return jsonify_result(db_helper.get_object_set())
-    elif request.method == 'POST':
-        if check_body(request, 'name'):
-            name = request.get_json()['name']
-            response = unwrap_db_result(db_helper.add_object_set(name))
-            response.status_code = 201;
-            return response
-        abort(400)
-
-@application.route('/api/object', methods=['GET', 'POST'])
-def objects():
-    if request.method == 'GET':
-        return jsonify_result(db_helper.get_object())
-    elif request.method == 'POST':
-        if check_body(request, 'name', 'object_set_id'):
-            name = request.get_json()['name']
-            object_set_id = request.get_json()['object_set_id']
-            response = unwrap_db_result(db_helper.add_object(name, object_set_id))
-            response.status_code = 201;
-            return response
-        abort(400)
-
-@application.route('/api/videos', methods=['GET', 'POST'])
-def videos():
-    if request.method == 'GET':
-        return jsonify_result(db_helper.get_video())
-    elif request.method == 'POST':
-        if check_body(request, 'file_path', 'camera_id'):
-            file_path = request.get_json()['file_path']
-            camera_id = request.get_json()['camera_id']
-            response = unwrap_db_result(db_helper.add_video(file_path, camera_id))
-            response.status_code = 201;
-            return response
-        abort(400)
-
 @application.route('/incident/<path:path>')
 def send_static_file(path):
     # Example: http://127.0.0.1:8000/incident/20210218/A210218_003304_003318.mp4
     return send_from_directory('incidents/converted', path)
+
+
+application.register_blueprint(api)
+db_helper.initialize(application)
 
 if __name__ == '__main__':
     # Fetch the latest incidents from the camera's FTP server
